Adds QuestionNotFoundError exception
Introduces a specific exception, `QuestionNotFoundError`, to handle cases where a question is not found in the repository.
This improves error handling and provides more context when a question cannot be retrieved, including the question ID, user ID, and collection name, if available.
It also removes the old exception from the repository specific location and moves it to the general exceptions location.
Removes try-except block from question provider and lets the exception bubble up.